move: BaseEntity 위치 변경 common/dto/BaseEntity -> common/entity/BaseEntity
