+++ src/main/java/salute/oneshot/domain/common/entity/BaseEntity.java
@@ -1,4 +1,4 @@
-package salute.oneshot.domain.common.dto.entity;
+package salute.oneshot.domain.common.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.EntityListeners;
