+++ src/main/java/salute/oneshot/domain/favorite/entity/Favorite.java
@@ -5,7 +5,7 @@ import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import salute.oneshot.domain.cocktail.entity.Cocktail;
-import salute.oneshot.domain.common.dto.entity.BaseEntity;
+import salute.oneshot.domain.common.entity.BaseEntity;
 
 @Entity
 @Table(name = "favorites", indexes = {
