+++ src/main/java/salute/oneshot/domain/recipeReview/entity/RecipeReview.java
@@ -12,7 +12,7 @@ import jakarta.persistence.Table;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import salute.oneshot.domain.cocktail.entity.Cocktail;
-import salute.oneshot.domain.common.dto.entity.BaseEntity;
+import salute.oneshot.domain.common.entity.BaseEntity;
 import salute.oneshot.domain.user.entity.User;
